pyproject.toml
[project]
name = "cross-efficient-vit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.8",
    "einops>=0.8.1",
    "facenet-pytorch>=2.6.0",
    "matplotlib>=3.10.5",
    "numpy>=1.26.4",
    "opencv-python-headless>=4.11.0.86",
    "pandas>=2.3.2",
    "pillow>=10.2.0",
    "progress>=1.6.1",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "termcolor>=3.1.0",
    "torch>=2.2.2",
    "torchaudio>=2.2.2",
    "torchvision>=0.17.2",
    "tqdm>=4.67.1",
]
